+++ pyroll/core/grooves/ovals/constricted_circular_oval.py
@@ -3,7 +3,7 @@ from typing import Optional
 import numpy as np
 
 from ..generic_elongation import GenericElongationGroove
-from ..utils import solve_r1234
+from ..generic_elongation_solvers import solve_r1234
 
 
 class ConstrictedCircularOvalGroove(GenericElongationGroove):
