+++ pyroll/core/grooves/ovals/oval_3radii.py
@@ -3,7 +3,7 @@ import numpy as np
 from scipy.optimize import minimize, Bounds
 from numpy import sin, cos, tan, pi, array
 from ..generic_elongation import GenericElongationGroove
-from ..utils import solve_r123
+from ..generic_elongation_solvers import solve_r123
 
 
 class Oval3RadiiGroove(GenericElongationGroove):
